+++ core/src/main/java/sting/ActAsStingComponent.java
@@ -24,6 +24,8 @@ import java.lang.annotation.Target;
  *
  * <p>It should be noted that Sting will attempt to use any annotation with this name and shape so that
  * frameworks do not need a direct code dependency on Sting.</p>
+ *
+ * @see InjectorFragment
  */
 @Documented
 @Retention( RetentionPolicy.RUNTIME )