+++ processor/src/main/java/sting/processor/Constants.java
@@ -11,6 +11,8 @@ final class Constants
   @Nonnull
   static final String INJECTOR_CLASSNAME = "sting.Injector";
   @Nonnull
+  static final String INJECTOR_FRAGMENT_CLASSNAME = "sting.InjectorFragment";
+  @Nonnull
   static final String INPUT_CLASSNAME = "sting.Injector.Input";
   @Nonnull
   static final String NAMED_CLASSNAME = "sting.Named";