+++ processor/src/test/fixtures/bad_input/com/example/injector_fragment/ClassInjectorFragment.java
@@ -1,0 +1,8 @@
+package com.example.injector_fragment;
+
+import sting.InjectorFragment;
+
+@InjectorFragment
+public class ClassInjectorFragment
+{
+}