+++ processor/src/test/fixtures/bad_input/com/example/injector_fragment/DefaultMethodInInjectorFragment.java
@@ -1,0 +1,11 @@
+package com.example.injector_fragment;
+
+import sting.InjectorFragment;
+
+@InjectorFragment
+public interface DefaultMethodInInjectorFragment
+{
+  default void doStuff()
+  {
+  }
+}