+++ processor/src/test/fixtures/input/com/example/named/NamedOnInjectorFragment.java
@@ -1,0 +1,11 @@
+package com.example.named;
+
+import sting.InjectorFragment;
+import sting.Named;
+
+@InjectorFragment
+interface NamedOnInjectorFragment
+{
+  @Named( "Foo" )
+  String getConfig();
+}